@@ -22,7 +22,13 @@
     "type": "git",
     "url": "git+ssh://git@github.com/hlerenow/imgToPixel.git"
   },
-  "keywords": ["image", "canvas", "imageToCanvas", "getImagePixel", "Pixel"],
+  "keywords": [
+    "image",
+    "canvas",
+    "imageToCanvas",
+    "getImagePixel",
+    "Pixel"
+  ],
   "dependencies": {},
   "deprecated": false,
   "description": "Convert image to pixel",
@@ -39,14 +45,16 @@
     "chalk": "^2.4.1",
     "cross-env": "^5.0.5",
     "glob": "^7.1.2",
-    "gulp": "^3.9.1",
+    "gulp": "^4.0.0",
     "gulp-babel": "^7.0.1",
+    "gulp-cli": "^2.0.1",
     "gulp-replace": "^1.0.0",
     "ora": "^3.0.0",
     "rimraf": "^2.6.2",
-    "webpack": "^3.6.0",
+    "webpack": "^4.28.4",
     "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-dev-server": "^2.9.1"
+    "webpack-cli": "^3.2.1",
+    "webpack-dev-server": "^3.1.14"
   },
   "license": "MIT"
 }
