@@ -1,12 +1,5 @@
-'use strict';
 
-var _main = require('./main.js');
-
-var _main2 = _interopRequireDefault(_main);
-
-function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
-
-;(function (name, definition) {
+;import ImgToPixel from './main.js';(function (name, definition) {
   // 检测上下文环境是否为AMD或CMD
   var hasDefine = typeof define === 'function',
 
@@ -24,5 +17,5 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
     window[name] = definition();
   }
 })('ImgToPixel', function () {
-  return _main2.default;
+  return ImgToPixel;
 });
